First batch of changes
Most of the changes I made are stylistic in nature. Mainly I wanted to implement more common techniques for accomplishing the same things. I think this script is very interesting but I would like to see it split up into a bunch of functions that handle different parts of the building process. I have implemented meta code like this before. It was helpful to have functions for handling the header, param block, begin block, process block, and end block. When you have a section of heavy weight code it is generally a good practice to wrap it up in a function. 

PowerShell uses a Controller and Tool framework structure. This means that we try to write tools to do specific things and controllers to orchestrate tools and controllers. 

PowerShell as a language does not have any line length. With that said people try to keep to about 110 characters as a somewhat hard limit and 80 as a soft. The second value is to keep it inline with python standards. I noticed a few lines that got very long. This slows the reading of your code down and makes it harder to reference back to when reading. Keeping to the soft standards helps speed up review on submission and in debugging

Over all I think this is an awesome idea and really want to see it mature =)